split carousel into 2 types
