--- frontend/src/components/Carousel/_Carousel.scss
@@ -1,49 +1,0 @@
-// Internal Imports
-@use "../Basics" as *;
-
-.click-carousel {
-  &.hidden {
-    @include hidden;
-  }
-
-  @include breakpoint-media-min("lgtablet") {
-    @include hidden;
-  }
-}
-
-.scroll-carousel {
-  display: flex;
-  //scroll-snap-type: x mandatory;
-  flex: 0 0 auto;
-
-  & > * {
-    //scroll-snap-align: center;
-    flex: 0 0 auto;
-    @supports (scroll-snap-stop: always) {
-      //scroll-snap-stop: always;
-    }
-  }
-
-  &.hidden {
-    @include hidden;
-  }
-
-  @include breakpoint-media-min("lgtablet") {
-    @include hidden;
-  }
-}
-
-//TODO delete once done
-.scroll-carousel-cop {
-  // matches small cop card
-  width: 100vw;
-  min-height: 550px;
-  overflow-x: hidden;
-  scrollbar-width: none;
-  justify-self: center;
-
-  & > * {
-    width: 85vw;
-    max-width: none;
-  }
-}
